Update-Database + Fix data

+++ Shop_Models/ViewModels/ProductDetailView.cs
@@ -4,8 +4,8 @@
     {
         public Guid Id { get; set; }
         public string? Ma { get; set; }
-        public decimal ImportPrice { get; set; }
-        public decimal Price { get; set; }
+        public double ImportPrice { get; set; }
+        public double Price { get; set; }
         public int AvailableQuantity { get; set; }
         public int Status { get; set; }
         public string? Description { get; set; }
